# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/files": {
            "post": {
                "description": "Сохраняет изображение в папке сервера images/. В поле \"img\" передается файл с расширениями jpg, png, webp, gif",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить изображение",
                "responses": {
                    "200": {
                        "description": "Цвет успешно создан",
                        "schema": {
                            "$ref": "#/definitions/views.SWGFileUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет файл. Требуется передовать только имя. Пример: example.gif",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Удалить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "название файла",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "файл успешно удалён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMessage"
                        }
                    },
                    "400": {
                        "description": "Неверное название",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "views.SWGError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "views.SWGFileUploadResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "example.jpg"
                }
            }
        },
...
        "views.SWGMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "some info"
                }
            }
        }
    }
}